+++ R/copula.R
@@ -463,7 +463,7 @@
 
 # copula filtering ---------------------------------------------------
 
-.copula_dynamic_filter <- function(object, y, update = TRUE, cond_mean = NULL, newxreg = NULL, ...)
+.copula_dynamic_filter <- function(object, y, update = TRUE, cond_mean = NULL, newxreg = NULL, newvreg = NULL, ...)
 {
     group <- NULL
     # the filtering will return a new estimate object with updated data.
@@ -476,6 +476,7 @@
     if (!is.xts(y)) stop("\ny must be an xts object.")
     .check_cond_mean_dynamics(object$spec$univariate, cond_mean)
     newxreg <- .multi_xreg_spec(object$spec$univariate, newxreg, NROW(y), "newxreg")
+    newvreg <- .multi_xreg_spec(object$spec$univariate, newvreg, NROW(y), "newvreg", slot = "vreg", missing_action = "error")
     if (!is.null(y)) {
         is_null_y <- FALSE
         new_y <- NROW(y)
@@ -497,7 +498,7 @@
     if (is.null(object$spec$target$original_size)) {
         object$spec$target$original_size <- NROW(object$spec$target$y)
     }
-    new_univariate <- .garch_filter_model(object, y, newxreg)
+    new_univariate <- .garch_filter_model(object, y, newxreg, newvreg)
     object$spec$target$y <- coredata(residuals(new_univariate))
     if (!is_null_y) {
         if (!is.null(cond_mean)) {
@@ -572,7 +573,7 @@
     return(object)
 }
 
-.copula_constant_filter <- function(object, y, update = TRUE, cond_mean = NULL, newxreg = NULL, ...)
+.copula_constant_filter <- function(object, y, update = TRUE, cond_mean = NULL, newxreg = NULL, newvreg = NULL, ...)
 {
 
     elapsed <- Sys.time()
@@ -580,6 +581,7 @@
     if (!is.xts(y)) stop("\ny must be an xts object.")
     .check_cond_mean_dynamics(object$spec$univariate, cond_mean)
     newxreg <- .multi_xreg_spec(object$spec$univariate, newxreg, NROW(y), "newxreg")
+    newvreg <- .multi_xreg_spec(object$spec$univariate, newvreg, NROW(y), "newvreg", slot = "vreg", missing_action = "error")
     if (!is.null(y)) {
         is_null_y <- FALSE
         new_y <- NROW(y)
@@ -600,7 +602,7 @@
     if (is.null(object$spec$target$original_size)) {
         object$spec$target$original_size <- NROW(object$spec$target$y)
     }
-    new_univariate <- .garch_filter_model(object, y, newxreg)
+    new_univariate <- .garch_filter_model(object, y, newxreg, newvreg)
     object$spec$target$y <- coredata(residuals(new_univariate))
     if (!is_null_y) {
         if (!is.null(cond_mean)) {
@@ -660,7 +662,7 @@
                                      Q_init = NULL, Z_init = NULL,
                                      init_method = c("start", "end"),
                                      cond_mean = NULL,
-                                     sim_method = c("parametric", "bootstrap"), xreg = NULL, ...)
+                                     sim_method = c("parametric", "bootstrap"), xreg = NULL, vreg = NULL, ...)
 {
     elapsed <- Sys.time()
     if (!is.null(seed)) set.seed(seed)
@@ -669,6 +671,7 @@
     .check_cond_mean_dynamics(object$spec$univariate, cond_mean)
     mu <- .cond_mean_spec(cond_mean, object$spec$n_series, h, object$spec$series_names)
     xreg <- .multi_xreg_spec(object$spec$univariate, xreg, h + burn, "xreg")
+    vreg <- .multi_xreg_spec(object$spec$univariate, vreg, h + burn, "vreg", slot = "vreg", missing_action = "error")
     group <- NULL
     Z <- object$copula_residuals
     R <- tscor(object)
@@ -730,7 +733,7 @@
     }
     gsim <- NULL
     gsim <- future_lapply(1:n_series, function(i){
-        .garch_simulate_model(object$spec$univariate[[i]], nsim, h - burn, burn, .retain_dimensions_array(std_residuals, i), init_method, xreg = xreg[[i]])
+        .garch_simulate_model(object$spec$univariate[[i]], nsim, h - burn, burn, .retain_dimensions_array(std_residuals, i), init_method, xreg = xreg[[i]], vreg = vreg[[i]])
     }, future.packages = c("tsgarch","tsmarch"), future.seed = TRUE)
     gsim <- eval(gsim)
     # reformat output [h n_series nsim]
@@ -765,7 +768,7 @@
     return(out)
 }
 
-.copula_constant_simulate_r <- function(object, nsim = 1, seed = NULL, h = 100, burn = 0, cond_mean = NULL, sim_method = "parametric", init_method = c("start", "end"), xreg = NULL, ...)
+.copula_constant_simulate_r <- function(object, nsim = 1, seed = NULL, h = 100, burn = 0, cond_mean = NULL, sim_method = "parametric", init_method = c("start", "end"), xreg = NULL, vreg = NULL, ...)
 {
     # simulate from t-copula
     parameter <- NULL
@@ -774,6 +777,7 @@
     .check_cond_mean_dynamics(object$spec$univariate, cond_mean)
     mu <- .cond_mean_spec(cond_mean, object$spec$n_series, h, object$spec$series_names)
     xreg <- .multi_xreg_spec(object$spec$univariate, xreg, h + burn, "xreg")
+    vreg <- .multi_xreg_spec(object$spec$univariate, vreg, h + burn, "vreg", slot = "vreg", missing_action = "error")
     init_method <- match.arg(init_method, c("start", "end"))
     h <- h + burn
     n_series <- object$spec$n_series
@@ -809,7 +813,7 @@
         std_residuals[,,i] <- .copula_qtransform(object, .retain_dimensions_array(U, i), i)
     }
     gsim <- lapply(1:n_series, function(i){
-        .garch_simulate_model(object$spec$univariate[[i]], nsim, h - burn, burn, .retain_dimensions_array(std_residuals, i), init_method, xreg = xreg[[i]])
+        .garch_simulate_model(object$spec$univariate[[i]], nsim, h - burn, burn, .retain_dimensions_array(std_residuals, i), init_method, xreg = xreg[[i]], vreg = vreg[[i]])
     })
     h <- h - burn
     sim_mu <- lapply(gsim, function(x) x$series)
@@ -842,7 +846,7 @@
 
 # copula prediction ---------------------------------------------------
 
-.copula_dynamic_predict <- function(object, h = 1, nsim = 1000, sim_method = c("parametric","bootstrap"), forc_dates = NULL, cond_mean = NULL, seed = NULL, newxreg = NULL, ...)
+.copula_dynamic_predict <- function(object, h = 1, nsim = 1000, sim_method = c("parametric","bootstrap"), forc_dates = NULL, cond_mean = NULL, seed = NULL, newxreg = NULL, newvreg = NULL, ...)
 {
     elapsed <- Sys.time()
     if (!is.null(seed)) set.seed(seed)
@@ -858,6 +862,7 @@
     .check_cond_mean_dynamics(object$spec$univariate, cond_mean)
     mu <- .cond_mean_spec(cond_mean, object$spec$n_series, h, object$spec$series_names)
     newxreg <- .multi_xreg_spec(object$spec$univariate, newxreg, h, "newxreg")
+    newvreg <- .multi_xreg_spec(object$spec$univariate, newvreg, h, "newvreg", slot = "vreg", missing_action = "error")
     group <- NULL
     Z <- object$copula_residuals
     R <- tscor(object)
@@ -917,7 +922,7 @@
     }
     gsim <- NULL
     gsim <- lapply(1:n_series, function(i){
-        .garch_simulate_model(object$spec$univariate[[i]], nsim, h, burn, .retain_dimensions_array(std_residuals, i), init_method, xreg = newxreg[[i]])
+        .garch_simulate_model(object$spec$univariate[[i]], nsim, h, burn, .retain_dimensions_array(std_residuals, i), init_method, xreg = newxreg[[i]], vreg = newvreg[[i]])
     })
     # reformat output [h n_series nsim]
     sim_mu <- lapply(gsim, function(x) x$series)
@@ -947,7 +952,7 @@
 }
 
 .copula_constant_predict <- function(object, h = 1, nsim = 1000, sim_method = c("parametric","bootstrap"), forc_dates = NULL,
-                                     cond_mean = NULL, seed = NULL, newxreg = NULL, ...)
+                                     cond_mean = NULL, seed = NULL, newxreg = NULL, newvreg = NULL, ...)
 {
     group <- NULL
     elapsed <- Sys.time()
@@ -965,6 +970,7 @@
     .check_cond_mean_dynamics(object$spec$univariate, cond_mean)
     mu <- .cond_mean_spec(cond_mean, object$spec$n_series, h, object$spec$series_names)
     newxreg <- .multi_xreg_spec(object$spec$univariate, newxreg, h, "newxreg")
+    newvreg <- .multi_xreg_spec(object$spec$univariate, newvreg, h, "newvreg", slot = "vreg", missing_action = "error")
     group <- NULL
     Z <- object$copula_residuals
     R <- tscor(object)
@@ -998,7 +1004,7 @@
         std_residuals[,,i] <- .copula_qtransform(object, .retain_dimensions_array(U, i), i)
     }
     gsim <- lapply(1:n_series, function(i){
-        .garch_simulate_model(object$spec$univariate[[i]], nsim, h, burn, .retain_dimensions_array(std_residuals, i), init_method, xreg = newxreg[[i]])
+        .garch_simulate_model(object$spec$univariate[[i]], nsim, h, burn, .retain_dimensions_array(std_residuals, i), init_method, xreg = newxreg[[i]], vreg = newvreg[[i]])
     })
     # reformat output [h n_series nsim]
     sim_mu <- lapply(gsim, function(x) x$series)