+++ R/dcc.R
@@ -285,7 +285,7 @@
 
 # dcc filtering ---------------------------------------------------
 
-.dcc_dynamic_filter <- function(object, y, update = TRUE, cond_mean = NULL, newxreg = NULL, ...)
+.dcc_dynamic_filter <- function(object, y, update = TRUE, cond_mean = NULL, newxreg = NULL, newvreg = NULL, ...)
 {
     group <- NULL
     # the filtering will return a new estimate object with updated data.
@@ -298,6 +298,7 @@
     if (!is.xts(y)) stop("\ny must be an xts object.")
     .check_cond_mean_dynamics(object$spec$univariate, cond_mean)
     newxreg <- .multi_xreg_spec(object$spec$univariate, newxreg, NROW(y), "newxreg")
+    newvreg <- .multi_xreg_spec(object$spec$univariate, newvreg, NROW(y), "newvreg", slot = "vreg", missing_action = "error")
     if (!is.null(y)) {
         is_null_y <- FALSE
         new_y <- NROW(y)
@@ -318,7 +319,7 @@
     if (is.null(object$spec$target$original_size)) {
         object$spec$target$original_size <- NROW(object$spec$target$y)
     }
-    new_univariate <- .garch_filter_model(object, y, newxreg)
+    new_univariate <- .garch_filter_model(object, y, newxreg, newvreg)
     object$spec$target$y <- coredata(residuals(new_univariate))
     if (!is_null_y) {
         if (!is.null(cond_mean)) {
@@ -387,7 +388,7 @@
     return(object)
 }
 
-.dcc_constant_filter <- function(object, y, update = TRUE, cond_mean = NULL, newxreg = NULL, ...)
+.dcc_constant_filter <- function(object, y, update = TRUE, cond_mean = NULL, newxreg = NULL, newvreg = NULL, ...)
 {
 
     elapsed <- Sys.time()
@@ -395,6 +396,7 @@
     if (!is.xts(y)) stop("\ny must be an xts object.")
     .check_cond_mean_dynamics(object$spec$univariate, cond_mean)
     newxreg <- .multi_xreg_spec(object$spec$univariate, newxreg, NROW(y), "newxreg")
+    newvreg <- .multi_xreg_spec(object$spec$univariate, newvreg, NROW(y), "newvreg", slot = "vreg", missing_action = "error")
     if (!is.null(y)) {
         is_null_y <- FALSE
         new_y <- NROW(y)
@@ -415,7 +417,7 @@
     if (is.null(object$spec$target$original_size)) {
         object$spec$target$original_size <- NROW(object$spec$target$y)
     }
-    new_univariate <- .garch_filter_model(object, y, newxreg)
+    new_univariate <- .garch_filter_model(object, y, newxreg, newvreg)
     object$spec$target$y <- coredata(residuals(new_univariate))
     if (!is_null_y) {
         if (!is.null(cond_mean)) {
@@ -466,7 +468,7 @@
 
 .dcc_dynamic_simulate_r <- function(object, nsim = 1, seed = NULL, h = 100, burn = 0,
                                     Q_init = NULL, Z_init = NULL, init_method = c("start", "end"),
-                                    cond_mean = NULL, sim_method = c("parametric", "bootstrap"), xreg = NULL, ...)
+                                    cond_mean = NULL, sim_method = c("parametric", "bootstrap"), xreg = NULL, vreg = NULL, ...)
 {
     elapsed <- Sys.time()
     if (!is.null(seed)) set.seed(seed)
@@ -476,6 +478,7 @@
     .check_cond_mean_dynamics(object$spec$univariate, cond_mean)
     mu <- .cond_mean_spec(cond_mean, object$spec$n_series, h, object$spec$series_names)
     xreg <- .multi_xreg_spec(object$spec$univariate, xreg, h + burn, "xreg")
+    vreg <- .multi_xreg_spec(object$spec$univariate, vreg, h + burn, "vreg", slot = "vreg", missing_action = "error")
     Z <- residuals(object, standardize = TRUE)
     R <- tscor(object)
     h <- h + burn
@@ -529,7 +532,7 @@
     R <- array(unlist(R_list), dim = c(h, vechn, nsim))
     gsim <- NULL
     gsim <- future_lapply(1:n_series, function(i){
-        .garch_simulate_model(object$spec$univariate[[i]], nsim, h - burn, burn, .retain_dimensions_array(std_residuals, i), init_method, xreg = xreg[[i]])
+        .garch_simulate_model(object$spec$univariate[[i]], nsim, h - burn, burn, .retain_dimensions_array(std_residuals, i), init_method, xreg = xreg[[i]], vreg = vreg[[i]])
     }, future.packages = c("tsgarch","tsmarch"), future.seed = TRUE)
     gsim <- eval(gsim)
     # reformat output [h n_series nsim]
@@ -564,7 +567,7 @@
     return(out)
 }
 
-.dcc_constant_simulate_r <- function(object, nsim = 1, seed = NULL, h = 100, burn = 0, cond_mean = NULL, init_method = c("start", "end"), xreg = NULL, ...)
+.dcc_constant_simulate_r <- function(object, nsim = 1, seed = NULL, h = 100, burn = 0, cond_mean = NULL, init_method = c("start", "end"), xreg = NULL, vreg = NULL, ...)
 {
     parameter <- NULL
     elapsed <- Sys.time()
@@ -572,6 +575,7 @@
     .check_cond_mean_dynamics(object$spec$univariate, cond_mean)
     mu <- .cond_mean_spec(cond_mean, object$spec$n_series, h, object$spec$series_names)
     xreg <- .multi_xreg_spec(object$spec$univariate, xreg, h + burn, "xreg")
+    vreg <- .multi_xreg_spec(object$spec$univariate, vreg, h + burn, "vreg", slot = "vreg", missing_action = "error")
     init_method <- match.arg(init_method, c("start", "end"))
     R <- object$R
     n_series <- object$spec$n_series
@@ -591,7 +595,7 @@
     }
     gsim <- NULL
     gsim <- future_lapply(1:n_series, function(i) {
-        .garch_simulate_model(object$spec$univariate[[i]], nsim, h, burn, .retain_dimensions_array(std_residuals, i), init_method, xreg = xreg[[i]])
+        .garch_simulate_model(object$spec$univariate[[i]], nsim, h, burn, .retain_dimensions_array(std_residuals, i), init_method, xreg = xreg[[i]], vreg = vreg[[i]])
     }, future.packages = c("tsgarch","tsmarch"), future.seed = TRUE, future.conditions = character(0L))
     gsim <- eval(gsim)
     # create covariance matrix
@@ -628,7 +632,7 @@
 # dcc prediction ---------------------------------------------------
 
 .dcc_dynamic_predict <- function(object, h = 1, nsim = 1000, sim_method = c("parametric","bootstrap"),
-                                 forc_dates = NULL, cond_mean = NULL, seed = NULL, newxreg = NULL, ...)
+                                 forc_dates = NULL, cond_mean = NULL, seed = NULL, newxreg = NULL, newvreg = NULL, ...)
 {
     parameter <- NULL
     elapsed <- Sys.time()
@@ -646,6 +650,7 @@
     .check_cond_mean_dynamics(object$spec$univariate, cond_mean)
     mu <- .cond_mean_spec(cond_mean, object$spec$n_series, h, object$spec$series_names)
     newxreg <- .multi_xreg_spec(object$spec$univariate, newxreg, h, "newxreg")
+    newvreg <- .multi_xreg_spec(object$spec$univariate, newvreg, h, "newvreg", slot = "vreg", missing_action = "error")
     Z <- residuals(object, standardize = TRUE)
     R <- tscor(object)
     alpha <- object$parmatrix[group == "alpha"]$value
@@ -700,7 +705,7 @@
     R <- array(unlist(R_list), dim = c(h, vechn, nsim))
     gsim <- NULL
     gsim <- lapply(1:n_series, function(i){
-        .garch_simulate_model(object$spec$univariate[[i]], nsim, h, burn, .retain_dimensions_array(std_residuals, i), init_method, xreg = newxreg[[i]])
+        .garch_simulate_model(object$spec$univariate[[i]], nsim, h, burn, .retain_dimensions_array(std_residuals, i), init_method, xreg = newxreg[[i]], vreg = newvreg[[i]])
     })
     # reformat output [h n_series nsim]
     sim_mu <- lapply(gsim, function(x) x$series)
@@ -732,7 +737,7 @@
 }
 
 .dcc_constant_predict <- function(object, h = 1, nsim = 1000, sim_method = c("parametric","bootstrap"), forc_dates = NULL,
-                                  cond_mean = NULL, seed = NULL, newxreg = NULL, ...)
+                                  cond_mean = NULL, seed = NULL, newxreg = NULL, newvreg = NULL, ...)
 {
     parameter <- group <- NULL
     elapsed <- Sys.time()
@@ -750,6 +755,7 @@
     .check_cond_mean_dynamics(object$spec$univariate, cond_mean)
     mu <- .cond_mean_spec(cond_mean, object$spec$n_series, h, object$spec$series_names)
     newxreg <- .multi_xreg_spec(object$spec$univariate, newxreg, h, "newxreg")
+    newvreg <- .multi_xreg_spec(object$spec$univariate, newvreg, h, "newvreg", slot = "vreg", missing_action = "error")
     group <- NULL
     Z <- residuals(object, standardize = TRUE)
     R <- tscor(object)
@@ -775,7 +781,7 @@
     std_residuals <- aperm(std_residuals, perm = c(3, 1, 2))
     # for U we need the array to be concentrated on n_series not nsim
     gsim <- lapply(1:n_series, function(i){
-        .garch_simulate_model(object$spec$univariate[[i]], nsim, h, burn, .retain_dimensions_array(std_residuals, i), init_method, xreg = newxreg[[i]])
+        .garch_simulate_model(object$spec$univariate[[i]], nsim, h, burn, .retain_dimensions_array(std_residuals, i), init_method, xreg = newxreg[[i]], vreg = newvreg[[i]])
     })
     # reformat output [h n_series nsim]
     sim_mu <- lapply(gsim, function(x) x$series)
@@ -894,7 +900,7 @@
 
 # dcc predict analytic ---------------------------------------------------
 
-.dcc_dynamic_predict_analytic <- function(object, h = 1, forc_dates = NULL, cond_mean = NULL, seed = NULL, newxreg = NULL, ...)
+.dcc_dynamic_predict_analytic <- function(object, h = 1, forc_dates = NULL, cond_mean = NULL, seed = NULL, newxreg = NULL, newvreg = NULL, ...)
 {
     elapsed <- Sys.time()
     if (!is.null(seed)) set.seed(seed)
@@ -909,6 +915,7 @@
     .check_cond_mean_dynamics(object$spec$univariate, cond_mean)
     mu <- .cond_mean_spec(cond_mean, object$spec$n_series, h, object$spec$series_names)
     newxreg <- .multi_xreg_spec(object$spec$univariate, newxreg, h, "newxreg")
+    newvreg <- .multi_xreg_spec(object$spec$univariate, newvreg, h, "newvreg", slot = "vreg", missing_action = "error")
     Z <- residuals(object, standardize = TRUE)
     R <- tscor(object)
     alpha <- object$parmatrix[group == "alpha"]$value
@@ -936,7 +943,7 @@
     exc <- maxpq
 
     garch_predictions <- lapply(1:n_series, function(i){
-        predict(object$spec$univariate[[i]], h = h, nsim  = 1, newxreg = newxreg[[i]])
+        predict(object$spec$univariate[[i]], h = h, nsim  = 1, newxreg = newxreg[[i]], newvreg = newvreg[[i]])
     })
 
     sigmas <- do.call(cbind, lapply(1:n_series, function(i) {