+++ R/univariate-fn.R
@@ -36,8 +36,9 @@
                 series_init = series_init, resid_init = resid_init))
 }
 
-.garch_simulate_model <- function(object, nsim, h, burn, innov, init_method, xreg = NULL)
+.garch_simulate_model <- function(object, nsim, h, burn, innov, init_method, xreg = NULL, vreg = NULL)
 {
+    group <- NULL
     init_states <- .garch_initialize_states(object, init_method)
     gspec <- object$spec
     gspec$parmatrix <- copy(object$parmatrix)
@@ -46,17 +47,21 @@
     if (!is.null(init_states$series_init)) args$series_init <- init_states$series_init
     if (!is.null(init_states$resid_init)) args$resid_init <- init_states$resid_init
     if (!is.null(xreg)) args$xreg <- xreg
+    if (!is.null(vreg)) {
+        xi <- object$parmatrix[group == "xi"]$value
+        args$vreg <- as.numeric(vreg %*% xi)
+    }
     sim <- do.call(simulate, args)
     return(sim)
 }
 
 
-.garch_filter_model <- function(object, y, newxreg = NULL)
+.garch_filter_model <- function(object, y, newxreg = NULL, newvreg = NULL)
 {
     m <- NCOL(y)
     new_fit <- NULL
     new_fit <- future_lapply(1:m, function(i) {
-        tsfilter(object$spec$univariate[[i]], y = y[,i], newxreg = newxreg[[i]])
+        tsfilter(object$spec$univariate[[i]], y = y[,i], newxreg = newxreg[[i]], newvreg = newvreg[[i]])
     }, future.packages = "tsgarch", future.seed = TRUE)
     new_fit <- eval(new_fit)
     names(new_fit) <- names(object$spec$univariate)